chore: extension 관련 Extension 디렉토리에 정리 fix: 커스텀 네비게이션 바 파라미터 추가
